requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
dnspython==2.8.0
filelock==3.20.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
python-whois==0.9.6
requests==2.32.5
requests-file==3.0.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tldextract==5.3.0
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
